🐛 fixing folder and file names and timestamp to educational institution
