+++ src/application/dtos/index.ts
@@ -1,5 +1,5 @@
 //Data Transfer Objects (interfaces para la entrada y salida de datos)
-export * from './educational_institution.dto';
+export * from './educationalInstitutionDto';
 export * from './major.dto';
 export * from './user.dto';
 export * from './jobOpportunity.dto';
