--- src/application/use-cases/educational-institution/create-educational-institution.use-case.ts
@@ -1,14 +1,0 @@
-import { EducationalInstitutionRepository } from '../../../infrastructure';
-import { EducationalInstitution } from '../../../domain';
-
-export class CreateEducationalInstitutionUseCase {
-  constructor(
-    private educationalInstitutionRepository: EducationalInstitutionRepository,
-  ) {}
-
-  public async execute(
-    data: Omit<EducationalInstitution, '_id'>,
-  ): Promise<EducationalInstitution> {
-    return this.educationalInstitutionRepository.create(data);
-  }
-}
