--- src/application/use-cases/educational-institution/delete-educational-institution.use-cases.ts
@@ -1,11 +1,0 @@
-import { EducationalInstitutionRepository } from '../../../infrastructure';
-
-export class DeleteEducationalInstitutionUseCase {
-  constructor(
-    private readonly educationalInstitutionRepository: EducationalInstitutionRepository,
-  ) {}
-
-  public async execute(id: string): Promise<boolean> {
-    return this.educationalInstitutionRepository.delete(id);
-  }
-}
