--- src/application/use-cases/educational-institution/get-all-educational-institution.use-case.ts
@@ -1,12 +1,0 @@
-import { EducationalInstitutionRepository } from '../../../infrastructure';
-import { EducationalInstitution } from '../../../domain';
-
-export class GetAllEducationalInstitutionUseCase {
-  constructor(
-    private readonly educationalInstitutionRepository: EducationalInstitutionRepository,
-  ) {}
-
-  public async execute(): Promise<EducationalInstitution[]> {
-    return this.educationalInstitutionRepository.findAll();
-  }
-}
