--- src/application/use-cases/educational-institution/get-educational-institution-by-id.use-case.ts
@@ -1,12 +1,0 @@
-import { EducationalInstitutionRepository } from '../../../infrastructure';
-import { EducationalInstitution } from '../../../domain';
-
-export class GetEducationalInstitutionByIdUseCase {
-  constructor(
-    private readonly educationalInstitutionRepository: EducationalInstitutionRepository,
-  ) {}
-
-  public async execute(id: string): Promise<EducationalInstitution | null> {
-    return this.educationalInstitutionRepository.findById(id);
-  }
-}
