+++ src/application/use-cases/educationalInstitution/getEducationalInstitutionByIdUseCase.ts
@@ -1,0 +1,12 @@
+import { EducationalInstitutionRepository } from '../../../infrastructure';
+import { EducationalInstitution } from '../../../domain';
+
+export class GetEducationalInstitutionByIdUseCase {
+  constructor(
+    private readonly educationalInstitutionRepository: EducationalInstitutionRepository,
+  ) {}
+
+  public async execute(id: string): Promise<EducationalInstitution | null> {
+    return this.educationalInstitutionRepository.findById(id);
+  }
+}
