+++ src/domain/entities/index.ts
@@ -1,6 +1,6 @@
 //Entidades de negocio puras
 //test
-export * from './educational-institution.entity';
+export * from './educationalInstitutionEntity';
 export * from './major.entity';
 export * from './user.entity';
 export * from './jobOpportunity.entity';
