+++ src/domain/repositories/index.ts
@@ -1,5 +1,5 @@
 //Interfaces de repositorios (sin implementación)
-export * from './educational-institution.repository';
+export * from './educationalInstitutionRepository';
 export * from './major.repository';
 export * from './user.repository';
 export * from './jobOpportunity.repository';
