+++ src/presentation/controllers/index.ts
@@ -1,5 +1,5 @@
 //Controladores que reciben la petición
-export * from './educational-institution.controller';
+export * from './educationalInstitutionController';
 export * from './major.controller';
 export * from './user.controller';
 export * from './jobOpportunity.controller';
