+++ src/presentation/routes/index.ts
@@ -1,5 +1,5 @@
 // Rutas de la API
-import educationalInstitutionRouter from './educational-institution.router';
+import educationalInstitutionRouter from './educationalInstitutionRouter';
 import majorRouter from './major.router';
 import userRouter from './user.router';
 import JobOpportunityRouter from './jobOpportunity.router';
